Moved sidebar into it's own separate component
